+++ vck-rqes/src/commonMain/kotlin/at/asitplus/wallet/lib/rqes/helper/OpenIdRqesParameters.kt
@@ -5,8 +5,7 @@ import at.asitplus.openid.TransactionData
 /**
  * Wrapper to distinguish RQES related [AuthenticationRequestParameter] members better
  */
-
-@Deprecated("Replaced", ReplaceWith("OpenIdRequestOptions"))
+@Deprecated("Subsumed", ReplaceWith("RqesRequestOptions"))
 data class OpenIdRqesParameters(
     val transactionData: Set<TransactionData>,
 )