+++ vck-rqes/src/commonTest/kotlin/at/asitplus/wallet/lib/rqes/RqesRequestOptionsTest.kt
@@ -5,28 +5,15 @@ import at.asitplus.openid.TransactionData
 import at.asitplus.rqes.QesInputDescriptor
 import at.asitplus.rqes.collection_entries.QesAuthorization
 import at.asitplus.rqes.collection_entries.RqesDocumentDigestEntry
-import at.asitplus.rqes.serializers.Base64URLTransactionDataSerializer
 import at.asitplus.signum.indispensable.Digest
-import at.asitplus.signum.indispensable.io.Base64UrlStrict
 import at.asitplus.wallet.eupid.EuPidScheme
 import at.asitplus.wallet.eupid.EuPidScheme.SdJwtAttributes.FAMILY_NAME
 import at.asitplus.wallet.eupid.EuPidScheme.SdJwtAttributes.GIVEN_NAME
-import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
-import at.asitplus.wallet.lib.agent.Holder
-import at.asitplus.wallet.lib.agent.HolderAgent
-import at.asitplus.wallet.lib.agent.IssuerAgent
-import at.asitplus.wallet.lib.agent.KeyMaterial
-import at.asitplus.wallet.lib.agent.toStoreCredentialInput
+import at.asitplus.wallet.lib.agent.*
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.SD_JWT
-import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.oidvci.encodeToParameters
-import at.asitplus.wallet.lib.openid.AuthenticationResponseResult
-import at.asitplus.wallet.lib.openid.AuthnResponseResult
-import at.asitplus.wallet.lib.openid.ClientIdScheme
-import at.asitplus.wallet.lib.openid.OpenId4VpHolder
+import at.asitplus.wallet.lib.openid.*
 import at.asitplus.wallet.lib.openid.OpenId4VpVerifier.CreationOptions.Query
-import at.asitplus.wallet.lib.openid.OpenIdRequestOptions
-import at.asitplus.wallet.lib.openid.RequestOptionsCredential
 import com.benasher44.uuid.bytes
 import com.benasher44.uuid.uuid4
 import io.kotest.core.spec.style.FreeSpec
@@ -36,10 +23,9 @@ import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldContain
 import io.kotest.matchers.types.shouldBeInstanceOf
 import io.ktor.http.*
-import io.matthewnelson.encoding.core.Decoder.Companion.decodeToByteArray
 
 @Suppress("DEPRECATION")
-class RqesOpenId4VpVerifierTest : FreeSpec({
+class RqesRequestOptionsTest : FreeSpec({
 
     lateinit var clientId: String
     lateinit var walletUrl: String
@@ -47,8 +33,8 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
     lateinit var verifierKeyMaterial: KeyMaterial
     lateinit var holderAgent: Holder
     lateinit var holderOid4vp: OpenId4VpHolder
-    lateinit var rqesVerifier: RqesOpenId4VpVerifier
-    lateinit var requestOptions: RqesOpenId4VpVerifier.ExtendedRequestOptions
+    lateinit var rqesVerifier: OpenId4VpVerifier
+    lateinit var requestOptions: RqesRequestOptions
 
     beforeEach {
         holderKeyMaterial = EphemeralKeyWithoutCert()
@@ -67,7 +53,7 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
         holderOid4vp = OpenId4VpHolder(
             holder = holderAgent,
         )
-        rqesVerifier = RqesOpenId4VpVerifier(
+        rqesVerifier = OpenId4VpVerifier(
             keyMaterial = verifierKeyMaterial,
             clientIdScheme = ClientIdScheme.RedirectUri(clientId),
         )
@@ -128,8 +114,8 @@ class RqesOpenId4VpVerifierTest : FreeSpec({
     }
 })
 
-private fun buildExtendedRequestOptions(): RqesOpenId4VpVerifier.ExtendedRequestOptions =
-    RqesOpenId4VpVerifier.ExtendedRequestOptions(
+private fun buildExtendedRequestOptions(): RqesRequestOptions =
+    RqesRequestOptions(
         baseRequestOptions = OpenIdRequestOptions(
             credentials = setOf(
                 RequestOptionsCredential(